明确io.edap.util.UnsafeUtil.writeByte方法调用Unsafe的方法签名

+++ edap-common/src/main/java/io/edap/util/UnsafeUtil.java
@@ -151,6 +151,6 @@ public class UnsafeUtil {
     }
 
     public static void writeByte(byte[] bs, int offset, byte b) {
-        UNSAFE.putByte(bs, UNSAFE.ARRAY_BYTE_BASE_OFFSET + offset, b);
+        UNSAFE.putByte(bs, (long)(UNSAFE.ARRAY_BYTE_BASE_OFFSET + offset), b);
     }
 }
